+++ src/components/packSelection/PackList.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>
+    List of packs...
+  </div>
+</template>
+
+<script>
+export default {
+
+};
+</script>
