+++ src/components/packSelection/PackSelection.vue
@@ -1,0 +1,19 @@
+<template>
+  <div>
+    <app-pack-list></app-pack-list>
+  </div>
+</template>
+
+<script>
+
+import appPackList from './PackList.vue';
+
+export default {
+  data() {
+    return {};
+  },
+  components: {
+    appPackList,
+  },
+};
+</script>
