+++ src/store/index.js
@@ -1,0 +1,11 @@
+import Vue from 'vue';
+import Vuex from 'vuex';
+
+Vue.use(Vuex);
+
+const store = new Vuex.Store({
+  modules: {
+  },
+});
+
+export { store as default };
